+++ packages/fastmcp_generator/CHANGELOG.md
@@ -1,3 +1,29 @@
+
+
+## 0.1.12
+
+- version bump to 0.1.12
+
+## 0.1.11
+
+- version bump to 0.1.11
+
+## 0.1.10
+
+- version bump to 0.1.10
+
+## 0.1.8
+
+- version bump to 0.1.8
+
+## 0.1.7
+
+- version bump to 0.1.7
+
+## 0.1.6
+
+- version bump to 0.1.6
+
 ## 0.1.5
 
 - version bump to 0.1.5